fiatRates: move coingecko-related code to a separate file, use interface
